+++ packages/http/src/index.ts
@@ -7,6 +7,8 @@ import route from './router';
 import { validateInput, validateOutput } from './validator';
 export * from './types';
 export * from './getHttpOperations';
+export * from './mocker/serializer/style';
+export { generate as generateHttpParam } from './mocker/generator/HttpParamGenerator';
 
 import { IHttpConfig, IHttpRequest, IHttpResponse, PickRequired, PrismHttpComponents } from './types';